detect pangram kata
level 6

Description:
A pangram is a sentence that contains every single letter of the alphabet at least once. For example, the sentence "The quick brown fox jumps over the lazy dog" is a pangram, because it uses the letters A-Z at least once (case is irrelevant).
Given a string, detect whether or not it is a pangram. Return True if it is, False if not. Ignore numbers and punctuation.

My Notes: need to create array of the alphabet, then take string. iterate through alphabet and check it against the string. cannot take string and check against the alphabet, since every letter is part of the alphabet.

My Solution: 
function isPangram(string){
	var alphabet = ['a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z'];
	var newString = string.toLowerCase().replace(/[^a-z]/g, "");

	for (var i=0; i< alphabet.length; i++) {
		var result = (newString.indexOf(alphabet[i])); 
		if (result === -1) {
			return false;
		}
		else {
			return true;
		}
	}
}